--- a/backend/package.json
+++ b/backend/package.json
@@ -10,14 +10,15 @@
     "scrape": "node scripts/scraper.js"
   },
   "dependencies": {
-    "express": "^4.18.2",
+    "axios": "^1.6.5",
+    "cheerio": "^1.0.0-rc.12",
     "cors": "^2.8.5",
-    "sqlite3": "^5.1.6",
+    "dotenv": "^16.3.1",
+    "express": "^4.18.2",
+    "express-rate-limit": "^7.5.1",
     "node-cron": "^3.0.3",
     "puppeteer": "^23.0.0",
-    "cheerio": "^1.0.0-rc.12",
-    "axios": "^1.6.5",
-    "dotenv": "^16.3.1"
+    "sqlite3": "^5.1.6"
   },
   "devDependencies": {
     "nodemon": "^3.0.2"