… backend-improvements.md
… backend/src/ScraperManager.js
… backend/src/scraper.js
… test-optimizations.js